fix: move rankCover to encodePassable
